% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ccp_programs}
\alias{ccp_programs}
\title{All current Career and College Promise (CCP) program codes.}
\format{
A data frame with 38 rows and 1 variable:
\describe{
  \item{Program}{Program code}
}
}
\usage{
ccp_programs
}
\description{
A dataset containing the program codes of all active and previous Career and College Promise (CCP) programs.
}
\keyword{datasets}
