% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.r
\name{coalesce_blanks}
\alias{coalesce_blanks}
\title{Coalesce a value to a non-null and non-blank.}
\usage{
coalesce_blanks(var, default = "")
}
\arguments{
\item{var}{Value to return if not null or blank.}

\item{default}{Value to return if var is null or blank.}
}
\description{
Coalesce a value to a non-null and non-blank.
}
