% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_description.r
\name{code_description}
\alias{code_description}
\title{Convert code to description}
\usage{
code_description(variable, code)
}
\arguments{
\item{variable}{Which variable to convert}

\item{code}{The coded value to convert}
}
\description{
Convert a code to its corresponding description.
}

