% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.r
\name{credential_seekers}
\alias{credential_seekers}
\title{Return a data frame of students who are curriculum credential seekers (seeking an Associate's, Diploma, or Certificate)}
\usage{
credential_seekers(report_year, report_semesters = NA_character_,
  exclude_hs = FALSE)
}
\arguments{
\item{report_year}{The year of the fall term for the data}

\item{report_semesters}{Either a single semester abbreviation or a list of semester abbreviations. If unspecified, all semesters are returned.}

\item{exlude_hs}{Should function exclude high school students from being included as credential seekers. Default is to include high school students.}
}
\description{
All data comes from IERG SQL Server database
}
