% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{early_college_programs}
\alias{early_college_programs}
\title{All current Early College (EC) program codes.}
\format{
A data frame with 3 rows and 1 variable:
\describe{
  \item{Program}{Program code}
}
}
\usage{
early_college_programs
}
\description{
A dataset containing the program codes of all active and previous Early College (EC) programs.
}
\keyword{datasets}
