% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.r
\name{fall_credential_seekers}
\alias{fall_credential_seekers}
\title{A special function to call credential_seekers for just a fall term}
\usage{
fall_credential_seekers(report_years, exclude_hs = FALSE)
}
\arguments{
\item{report_years}{The year of the fall term for the data}

\item{exlude_hs}{Should function exclude high school students from being included as credential seekers. Default is to include high school students.}
}
\description{
All data comes from IERG SQL Server database
}
