% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.r
\name{fall_enrollment}
\alias{fall_enrollment}
\title{A special function to call term_enrollment for just a fall term}
\usage{
fall_enrollment(report_years = NA)
}
\arguments{
\item{report_years}{The year of the fall term for the data}
}
\description{
All data comes from CCDW_HIST SQL Server database
}
