% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informer.r, R/test_do_not_src_control.r
\name{getCfg}
\alias{getCfg}
\title{Load the YAML configuration file.}
\usage{
getCfg(cfg_fn = "config.yml", cfg_path = ".", reload = FALSE)

getCfg(cfg_fn = "config.yml", cfg_path = ".", reload = FALSE)
}
\arguments{
\item{cfg_fn}{The file name for the YAML configuration file. Defaults to config.yml.}

\item{cfg_path}{The file path to the YAML configuration file. Defaults to ".".}

\item{reload}{Force a reload of the config.}
}
\description{
Load the specified YAML configuration file. If no file is provided,
load "./config.yml".

Return a data from of the IPEDS cohort data. Data will come either from the file ipeds_cohorts.csv of
from the CCDW_HIST SQL Server database.
}
