% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{\%like any\%}
\alias{\%like any\%}
\title{Return an element if it is like any one of the provided patterns.}
\usage{
x \%like any\% pattern
}
\arguments{
\item{x}{Elements to return if it matches any one pattern.}

\item{pattern}{Patterns to match.}
}
\description{
Return an element if it is like any one of the provided patterns.
}
