% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.r
\name{\%like\%}
\alias{\%like\%}
\title{Just another name for `%like any%`.}
\usage{
x \%like\% pattern
}
\arguments{
\item{x}{Elements to return if it matches any one pattern.}

\item{pattern}{Patterns to match.}
}
\description{
Just another name for `%like any%`.
}
