% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{\%nin\%}
\alias{\%nin\%}
\title{Return a list of elements from x that are not in y.}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{Elements that are not to be found in the List y.}

\item{y}{List to check against.}
}
\description{
Return a list of elements from x that are not in y.
}
