% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local.r
\name{\%nlike\%}
\alias{\%nlike\%}
\title{The opposite of `%like any%`.}
\usage{
x \%nlike\% pattern
}
\arguments{
\item{x}{Elements to return if it matches any one pattern.}

\item{pattern}{Patterns to match.}
}
\description{
The opposite of `%like any%`.
}
