% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{\%||\%}
\alias{\%||\%}
\title{If lhs is null, then return rhs.}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{Element to return if it is not null.}

\item{rhs}{Element to return if lhs is null.}
}
\description{
If lhs is null, then return rhs.
}
