% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{haywood_county_high_schools}
\alias{haywood_county_high_schools}
\title{All current high schools and codes for Haywood County.}
\format{A data frame with 4 rows and 2 variable:
\describe{
  \item{School}{School Name}
  \item{Institution_ID}{The Colleague institution ID}
}}
\usage{
haywood_county_high_schools
}
\description{
A dataset containing the high school names and codes for Haywood County.
}
\keyword{datasets}
