% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.r
\name{high_school_graduation_dates}
\alias{high_school_graduation_dates}
\title{Load the YAML configuration file.}
\usage{
high_school_graduation_dates()
}
\description{
Load the specified YAML configuration file. If no file is provided,
load "./config.yml".
}
