% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{high_school_programs}
\alias{high_school_programs}
\title{All current high school program codes.}
\format{
A data frame with 41 rows and 1 variable:
\describe{
  \item{Program}{Program code}
}
}
\usage{
high_school_programs
}
\description{
A dataset containing the program codes of all active and previous high school programs.
}
\keyword{datasets}
