% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.r
\name{ipeds_cohort}
\alias{ipeds_cohort}
\title{Return a data from of the IPEDS cohort data.}
\usage{
ipeds_cohort(report_year, cohorts = c("FT", "PT", "TF", "TP", "RF",
  "RP"))
}
\arguments{
\item{report_year}{The year of the fall term for the data}

\item{cohorts}{Which cohorts to include in data frame. FT = Full-time First-time, PT = Part-time First-time,
TF = Full-time Transfer, TP = Part-time Transfer,
RF = Full-time Returning, RP = Part-time Returning}
}
\description{
Return a data from of the IPEDS cohort data. Data will come either from the file ipeds_cohorts.csv of
from the IERG SQL Server database.
}
