% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.r
\name{merge_files}
\alias{merge_files}
\title{Merge all files in a specified path}
\usage{
merge_files(path = getwd(), pattern = "*", type = "csv", FNvar = "",
  col_types = cols(.default = col_character()), ...)
}
\arguments{
\item{path}{Defaults to the current working directory using getwd()}

\item{pattern}{The pattern used to filter the files prior to the merge}

\item{type}{The type of merge. Possible options are csv, fixed, other}

\item{FNvar}{The name of the variable which will contain the filename. Will only add filename if this parameter has a variable name.}
}
\description{
This function allows you to merge all the files in a specified folder. 
It handles removing the extra header lines, if necessary. It also 
adds the file name as an extra variable, if this is necessary.
}
\keyword{file}

