% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/term_code.r
\name{term_code}
\alias{term_code}
\title{Convert term sequence to term code}
\usage{
term_code(termseq)
}
\arguments{
\item{termseq}{A term sequence string - YYYYNN where NN is 01, 02, or 03}
}
\description{
Convert term sequence to term code.
}

