% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.r
\name{credential_seekers}
\alias{credential_seekers}
\title{Return a data frame of students who are curriculum credential seekers (seeking an Associate's, Diploma, or Certificate)}
\usage{
credential_seekers(
  report_years = NA_integer_,
  report_semesters = NA_character_,
  exclude_hs = FALSE
)
}
\arguments{
\item{report_years}{The year or a list of years of the fall term for the data. If unspecified, all years are returned.}

\item{report_semesters}{Either a single semester abbreviation or a list of semester abbreviations. If unspecified, all semesters are returned.}

\item{exclude_hs}{Should function exclude high school students from being included as credential seekers. Default is to include high school students.}
}
\description{
All data comes from CCDW_HIST SQL Server database
}
