% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/informer.r
\name{getCfg}
\alias{getCfg}
\title{Load the YAML configuration file.}
\usage{
getCfg(
  cfg_full_path = NA_character_,
  cfg_fn = NA_character_,
  cfg_path = NA_character_,
  reload = FALSE
)
}
\arguments{
\item{cfg_full_path}{Provide the full path to the YAML configuration file. Defaults to NA.}

\item{cfg_fn}{The file name for the YAML configuration file. Defaults to config.yml.}

\item{cfg_path}{The file path to the YAML configuration file. Defaults to ".".}

\item{reload}{Force a reload of the config.}
}
\description{
Load the specified YAML configuration file. If no file is provided,
load "./config.yml".
}
