% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipeds.r
\name{ipeds_cohort}
\alias{ipeds_cohort}
\title{Return a data from of the IPEDS cohort data.}
\usage{
ipeds_cohort(
  report_years,
  cohorts = c("FT", "PT", "TF", "TP", "RF", "RP"),
  use = "ipeds_cohorts",
  ipeds_path,
  useonly = TRUE
)
}
\arguments{
\item{report_years}{The year of the fall term for the data}

\item{cohorts}{Which cohorts to include in data frame. FT = Full-time First-time, PT = Part-time First-time,
TF = Full-time Transfer, TP = Part-time Transfer,
RF = Full-time Returning, RP = Part-time Returning}

\item{use}{Which dataset should be used for cohorts from Colleague
ipeds_cohorts Use the local.ipeds_cohorts table
STUDENT_TERMS Use the history.STUDENT_TERMS_Current view}

\item{ipeds_path}{The path where ipeds_cohort.csv file is located.}

\item{useonly}{Use the database cohort found in the table specified in the
`use` parameter only. Default is FALSE which means combine data from
the database table with the file ipeds_cohorts.csv.}
}
\description{
Return a data from of the IPEDS cohort data. Data will come either from the file ipeds_cohorts.csv or
from the CCDW_HIST SQL Server database.
}
